requirements.txt
django
pillow
slippers
django-browser-reload
kolo
razorpay
matplotlib
django-plotly-dash
# plotly
# pandas
# kaleido